test: Participant 테스트 로직 PariticipantTest로 이동
