+++ src/test/java/blackJack/domain/participant/PlayerTest.java
@@ -58,61 +58,6 @@ class PlayerTest {
     }
 
     @Test
-    @DisplayName("플레이어의 카드 합계 계산 테스트")
-    void calculateScore() {
-        Player player = new Player("rookie");
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.EIGHT));
-
-        assertThat(player.getScore()).isEqualTo(8);
-    }
-
-    @Test
-    @DisplayName("플레이어의 카드에 Ace가 11로 되는 경우 합계 계산 테스트")
-    void calculateScoreWithAceEleven() {
-        Player player = new Player("rookie");
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.ACE));
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.JACK));
-
-        assertThat(player.getScore()).isEqualTo(21);
-    }
-
-    @Test
-    @DisplayName("플레이어의 카드에 Ace가 1로 되는 경우 합계 계산 테스트")
-    void calculateScoreWithAceOne() {
-        Player player = new Player("rookie");
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.ACE));
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.JACK));
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.EIGHT));
-
-        assertThat(player.getScore()).isEqualTo(19);
-    }
-
-    @Test
-    @DisplayName("플레이어의 카드에 Ace가 여러개인 경우 계산 테스트")
-    void calculateScoreWithAceCountThree() {
-        Player player = new Player("rookie");
-        player.receiveCard(Card.from(Symbol.CLOVER, Denomination.ACE));
-        player.receiveCard(Card.from(Symbol.HEART, Denomination.ACE));
-        player.receiveCard(Card.from(Symbol.DIAMOND, Denomination.ACE));
-        player.receiveCard(Card.from(Symbol.SPADE, Denomination.EIGHT));
-
-        assertThat(player.getScore()).isEqualTo(21);
-    }
-
-    @Test
-    @DisplayName("중복된 카드를 받는 경우 예외 발생 테스트")
-    void receiveDuplicatedCard() {
-        Card card1 = Card.from(Symbol.CLOVER, Denomination.ACE);
-        Card card2 = Card.from(Symbol.CLOVER, Denomination.ACE);
-        Participant participant = new Player("rookie");
-        participant.receiveCard(card1);
-
-        assertThatThrownBy(() -> participant.receiveCard(card2))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("중복된 카드는 받을 수 없습니다.");
-    }
-
-    @Test
     @DisplayName("플레이어가 승리하는 경우 승부 결과 반환 테스트")
     void getMatchResultPlayerWin() {
         Player player = new Player("rookie");
